backend/requirements.txt
Django>=5.1,<6.0
djangorestframework==3.15.*
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
django-celery-beat>=2.6,<3.0
psycopg[binary]>=3.1,<4.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
openpyxl>=3.1,<4.0
gunicorn>=22.0,<23.0
python-decouple>=3.8
